bgpd: move RTC code to bgp_rtc module
A little late, but ... move the RTC feature code to its
own module.
