+++ bgpd/bgp_route.h
@@ -827,6 +827,7 @@ extern void bgp_redistribute_delete(struct bgp *, struct prefix *, uint8_t,
 				    unsigned short);
 extern void bgp_redistribute_withdraw(struct bgp *, afi_t, int, unsigned short);
 
+struct bgp_static *bgp_static_new(void);
 extern void bgp_static_add(struct bgp *);
 extern void bgp_static_delete(struct bgp *);
 extern void bgp_address_family_distance_delete(void);
@@ -1015,30 +1016,5 @@ extern int early_route_process(struct bgp *bgp, struct bgp_dest *dest);
 extern int other_route_process(struct bgp *bgp, struct bgp_dest *dest);
 extern int eoiu_marker_process(struct bgp *bgp, struct bgp_dest *dest);
 extern uint32_t bgp_med_value(struct attr *attr, struct bgp *bgp);
-
-/* Max prefixlen for an RTC "prefix", formed from an RT */
-#define BGP_RTC_PREFIX_MAXLEN 96
-
-/* Init, deinit peer's RTC data */
-void bgp_rtc_peer_init(struct peer *peer);
-void bgp_rtc_peer_delete(struct peer *peer);
-/* RTC import list change: may be "remove imports", or "add/change imports" */
-int bgp_rtc_import_update(struct bgp *bgp, const struct ecommunity *oldcomm,
-			  const struct ecommunity *newcomm, bool update);
-/* RTC peer activate/deactivate change */
-int bgp_rtc_peer_update(struct peer *peer, afi_t afi, safi_t safi, bool active);
-/* Finer-grained RT change for RTC: add or remove one RT */
-struct ecommunity_val; /* Forward ref */
-int bgp_rtc_import_change(struct bgp *bgp, const struct ecommunity_val *eval,
-			  bool add_p);
-/* RTC prefix advertisement update */
-int bgp_rtc_prefix_update(struct bgp_dest *dest, struct bgp_path_info *oldpi,
-			  struct bgp_path_info *newpi);
-/* Special handling for peer advertising the RTC default prefix */
-int bgp_rtc_default_update(struct peer *peer, const struct prefix *p,
-			   bool add_p);
-/* Show output helper */
-void bgp_rtc_show_peer(const struct peer *peer, struct vty *vty,
-		       json_object *jneigh);
 
 #endif /* _QUAGGA_BGP_ROUTE_H */
